// game/CDA/CognitiveReaction/src/interface.d.ts
export interface ICreateParams {
  /** CreateParams exchangeRate */
  exchangeRate?: number | null;

  /** CreateParams timeLimit */
  timeLimit?: number | null;

  /** CreateParams nextPhaseKey */
  nextPhaseKey?: string | null;
}
// ...
export interface IGameState {
  /** GameState gameStage */
  gameStage?: number | null;

  /** GameState time */
  time?: number | null;
}
// ...
export interface IPlayerState {
  /** PlayerState seatNumber */
  seatNumber?: number | null;

  /** PlayerState correctNumber */
  correctNumber?: number | null;

  /** PlayerState point */
  point?: number | null;

  /** PlayerState answers */
  answers?: string[] | null;

  /** PlayerState index */
  index?: number | null;

  /** PlayerState playerStage */
  playerStage?: number | null;
}
// ...
export interface IMoveParams {
  /** MoveParams seatNumber */
  seatNumber?: number | null;

  /** MoveParams answer */
  answer?: string | null;
}
// ...
export interface IPushParams {}